pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "audio-monitor"
version = "0.1.0"
description = "Lightweight balcony audio monitoring MVP for Raspberry Pi 3 B+"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "Flask>=3.0,<4.0",
  "numpy>=1.26,<3.0",
  "PyYAML>=6.0,<7.0",
  "waitress>=3.0,<4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

